Data: refactored state objects
