+++ webapp/loggedin/modules/data/dataVis/dataQuery/actions/fetch.js
@@ -2,7 +2,7 @@ import axios from 'axios'
 
 import * as DataSort from '@common/surveyRdb/dataSort'
 import * as SurveyState from '@webapp/survey/surveyState'
-import * as DataQueryState from '@webapp/loggedin/modules/data/dataVis/dataQuery/dataQueryState'
+import * as DataQueryState from '@webapp/loggedin/modules/data/dataVis/dataQuery/state'
 
 export const dataQueryTableInit = 'dataQuery/table/init'
