+++ webapp/loggedin/modules/data/dataVis/dataQuery/state/defaults.js
@@ -1,0 +1,8 @@
+export const defaults = {
+  offset: 0,
+  limit: 15,
+  data: [],
+  filter: null,
+  nodeDefUuidCols: [],
+  sort: [],
+}
