+++ webapp/loggedin/modules/data/dataVis/dataQuery/state/index.js
@@ -1,0 +1,26 @@
+export { defaults } from './defaults'
+
+export {
+  assocNodeDefUuidCols,
+  assocNodeDefUuidTable,
+  assocTableData,
+  assocTableDataCol,
+  assocTableFilter,
+  assocTableSort,
+  dissocTableDataCols,
+  getTableCount,
+  getTableData,
+  getTableEditMode,
+  getTableFilter,
+  getTableLimit,
+  getTableNodeDefUuidCols,
+  getTableNodeDefUuidTable,
+  getTableOffset,
+  getTableSort,
+  hasTableAndCols,
+  initTableData,
+} from './table'
+
+export { assocShowNodeDefSelectors, isNodeDefSelectorsVisible } from './nodeDefSelectors'
+
+export { assocTableDataRecordNodes, assocTableDataRecordNodeValidations } from './node'
