+++ webapp/loggedin/modules/data/dataVis/dataQuery/state/keys.js
@@ -1,0 +1,20 @@
+export const keys = {
+  table: 'table',
+  showNodeDefSelectors: 'showNodeDefSelectors',
+}
+
+export const tableKeys = {
+  data: 'data',
+  offset: 'offset',
+  limit: 'limit',
+  count: 'count',
+  filter: 'filter',
+  sort: 'sort',
+  nodeDefUuidTable: 'nodeDefUuidTable',
+  nodeDefUuidCols: 'nodeDefUuidCols',
+  editMode: 'editMode',
+}
+
+export const rowKeys = {
+  record: 'record',
+}
