+++ webapp/loggedin/modules/data/dataVis/dataQuery/state/read.js
@@ -1,0 +1,4 @@
+import * as R from 'ramda'
+import * as DataVisState from '@webapp/loggedin/modules/data/dataVis/state'
+
+export const getState = R.pipe(DataVisState.getState, R.prop('query'))
