+++ webapp/loggedin/modules/data/dataVis/dataQuery/state/table.js
@@ -1,43 +1,10 @@
 import * as R from 'ramda'
 
-import * as Record from '@core/record/record'
-import * as Node from '@core/record/node'
-import * as Validation from '@core/validation/validation'
-import * as DataVisState from '../dataVisState'
+import { getState } from './read'
+import { defaults } from './defaults'
+import { keys, tableKeys } from './keys'
 
-export const defaults = {
-  offset: 0,
-  limit: 15,
-  data: [],
-  filter: null,
-  nodeDefUuidCols: [],
-  sort: [],
-}
-
-const getState = R.pipe(DataVisState.getState, R.prop('query'))
-
-const keys = {
-  table: 'table',
-  showNodeDefSelectors: 'showNodeDefSelectors',
-}
-
-const tableKeys = {
-  data: 'data',
-  offset: 'offset',
-  limit: 'limit',
-  count: 'count',
-  filter: 'filter',
-  sort: 'sort',
-  nodeDefUuidTable: 'nodeDefUuidTable',
-  nodeDefUuidCols: 'nodeDefUuidCols',
-  editMode: 'editMode',
-}
-
-const rowKeys = {
-  record: 'record',
-}
-
-// ====== table
+// ====== READ
 const getTableProp = (tableProp, defaultValue = null) =>
   R.pipe(getState, R.pathOr(defaultValue, [keys.table, tableProp]))
 
@@ -55,6 +22,7 @@ const hasTable = R.pipe(getTableNodeDefUuidTable, R.isNil, R.not)
 const hasCols = R.pipe(getTableNodeDefUuidCols, R.isEmpty, R.not)
 export const hasTableAndCols = (state) => hasTable(state) && hasCols(state)
 
+// ====== UPDATE
 export const assocTableData = (offset, data) =>
   R.pipe(R.assocPath([keys.table, tableKeys.offset], offset), R.assocPath([keys.table, tableKeys.data], data))
 
@@ -101,53 +69,3 @@ export const initTableData = (offset, limit, filter, sort, count, data, nodeDefU
     nodeDefUuidCols,
     editMode,
   })
-
-// ===== Edit mode
-const _updateData = (updateFn) => (state) =>
-  R.pipe(R.pathOr([], [keys.table, tableKeys.data]), updateFn, (dataUpdated) =>
-    R.assocPath([keys.table, tableKeys.data], dataUpdated, state)
-  )(state)
-
-export const assocTableDataRecordNodes = (nodes) =>
-  // Replace nodes in table rows
-  _updateData((data) =>
-    Object.values(nodes).reduce(
-      (accData, node) =>
-        accData.map((row) => {
-          const nodeDefUuid = Node.getNodeDefUuid(node)
-          const rowUpdated = { ...row }
-          const { record } = rowUpdated
-          if (Record.getUuid(record) === Node.getRecordUuid(node)) {
-            const cell = rowUpdated.cols[nodeDefUuid]
-
-            if (cell && Node.isEqual(node)(cell.node)) {
-              rowUpdated.cols[nodeDefUuid] = R.ifElse(
-                R.always(Node.isDeleted(node)),
-                R.dissoc('node'),
-                R.assoc('node', node)
-              )(cell)
-            }
-          }
-          return rowUpdated
-        }),
-      data
-    )
-  )
-
-export const assocTableDataRecordNodeValidations = (recordUuid, validations) =>
-  _updateData(
-    R.map((row) => {
-      if (R.pathEq([rowKeys.record, Record.keys.uuid], recordUuid)(row)) {
-        const record = row[rowKeys.record]
-        const recordValidation = Validation.getValidation(record)
-        const recordValidationUpdated = Validation.mergeValidation(validations)(recordValidation)
-        return R.assocPath([rowKeys.record, Validation.keys.validation], recordValidationUpdated)(row)
-      }
-      return row
-    })
-  )
-
-// ====== nodeDefSelectors
-
-export const isNodeDefSelectorsVisible = R.pipe(getState, R.propOr(true, keys.showNodeDefSelectors))
-export const assocShowNodeDefSelectors = R.assoc(keys.showNodeDefSelectors)
