+++ webapp/loggedin/modules/data/dataVis/state/index.js
@@ -1,4 +1,4 @@
 import * as R from 'ramda'
-import * as DataState from '../dataState'
+import * as DataState from '@webapp/loggedin/modules/data/state'
 
 export const getState = R.pipe(DataState.getState, R.prop('dataVis'))
